requirements.txt
nicegui>=1.4.0
watchdog>=3.0.0
tree-sitter>=0.20.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pyyaml>=6.0
pydantic>=2.0.0
jsonschema>=4.17.0
sqlite-utils>=3.34
pytest>=7.4.0
